Port to MinGW
Compile-tested only.

+++ lib/init.c
@@ -293,7 +293,8 @@ iscsi_create_context(const char *initiator_name)
 	while (iscsi->smalloc_size < required) {
 		iscsi->smalloc_size <<= 1;
 	}
-	ISCSI_LOG(iscsi,5,"small allocation size is %zu byte", iscsi->smalloc_size);
+	ISCSI_LOG(iscsi, 5, "small allocation size is %u byte",
+                  (uint32_t)iscsi->smalloc_size);
 
 	ca = getenv("LIBISCSI_CACHE_ALLOCATIONS");
 	if (!ca || atoi(ca) != 0) {
